requirements.txt
fastapi
uvicorn
logging
OPi.GPIO